Implementing through external_download

+++ R/download.R
@@ -274,7 +274,8 @@ sidra_download <- function(sidra_code = NULL, year, geo_level = "municipality",
   } # End of If - Download at the Municipality Level
 }
 
-external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_level = NULL) {
+external_download <- function(dataset = NULL, source = NULL, year = NULL,
+                              geo_level = NULL, coords = NULL, dataset_code = NULL) {
 
   ## Bind GLobal Variables
 
@@ -294,6 +295,8 @@ external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_le
   param$source <- source
   param$year <- year # This may not make sense if data is downloaded for all time periods together
   param$geo_level <- geo_level # This could also not make sense
+  param$coords <- coords
+  param$dataset_code <- dataset_code
 
   # if (param$geo_level == 'legal_amazon' & param$source == 'prodes'){param$geo_level = 'legal-amz-prodes'}
   # if (param$geo_level == 'amazon_biome' & param$source == 'prodes'){param$geo_level = 'amz-prodes'}
@@ -451,6 +454,45 @@ external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_le
     }
   }
 
+  ##################
+  ## TerraClimate ##
+  ##################
+
+  if (source == "terraclimate"){
+
+    filename <- paste0(
+      "agg_terraclimate_",
+      param$dataset_code,
+      "_1958_CurrentYear_GLOBE.nc"
+    )
+
+    path <- paste0(
+      param$url,
+      "/",
+      filename,
+      "?",
+      "&var=",
+      param$dataset_code,
+
+      "&south=",
+      param$coords$lat_min,
+      "&north=",
+      param$coords$lat_max,
+      "&west=",
+      param$coords$lon_min,
+      "&east=",
+      param$coords$lon_max,
+      "&horizStride=1",
+
+      "&time_start=",
+      param$year$initial_time,
+      "&time_end=",
+      param$year$final_time,
+      "&timeStride=1",
+
+      "&disableProjSubset=on&addLatLon=true&accept=netcdf"
+    )
+  }
 
   #######################
   ## Initiate Download ##
@@ -477,6 +519,9 @@ external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_le
   if (source == "ibama") {
     file_extension <- ".zip"
   }
+  if (source == "terraclimate"){
+    file_extension <- ".nc"
+  }
 
   # !!!  We should Change This to a Curl Process
 
@@ -488,7 +533,7 @@ external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_le
   ## Extraction through Curl Requests
   ## Investigate a bit more on how Curl Requests work
 
-  if (!(source %in% c("deter", "seeg"))) {
+  if (!(source %in% c("deter", "seeg", "terraclimate"))) {
     utils::download.file(url = path, destfile = temp, mode = "wb")
   }
   if (source == "deter") {
@@ -503,6 +548,10 @@ external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_le
     if (geo_level == "municipality") {
       googledrive::drive_download(path, path = temp, overwrite = TRUE)
     }
+  }
+  if (source == "terraclimate"){
+    utils::download.file(url = path, destfile = temp, method = "curl"
+    )
   }
 
   ## This Data Opening Part Should Include the .Shp, not DBF
@@ -526,6 +575,9 @@ external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_le
     dat <- readr::read_csv(temp, locale = readr::locale(encoding = "latin1")) %>%
       janitor::clean_names() %>%
       tibble::as_tibble()
+  }
+  if (file_extension == ".nc"){
+    dat <- terra::rast(temp)
   }
   if (file_extension == ".xlsx") {
     if (param$dataset == "mapbiomas_cover") {
@@ -659,7 +711,9 @@ external_download <- function(dataset = NULL, source = NULL, year = NULL, geo_le
 
   # Folder is kept
 
-  unlink(temp)
+  if (file_extension != ".nc") {
+    unlink(temp)
+    }
 
   ####################
   ## Pre-Processing ##
@@ -868,6 +922,33 @@ datasets_link <- function() {
     ## Demographic Info ##
 
     # https://sidra.ibge.gov.br/pesquisa/censo-demografico/series-temporais/series-temporais/
+
+    ##########
+    ## CIPÓ ##
+    ##########
+
+    'CIPO', 'brazilian_actors', NA, NA, NA, 'https://plataformacipo.org/mapa-crimes-ambientais/',
+    'CIPO', 'international_cooperation', NA, NA, NA, 'https://plataformacipo.org/mapeamento-cooperacao-internacional/',
+    'CIPO', 'forest_governance', NA, NA, NA, 'https://plataformacipo.org/arranjos-globais-de-governanca-florestal/',
+
+    ##################
+    ## TerraClimate ##
+    ##################
+
+    'TerraClimate', 'max_temperature', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'min_temperature', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'wind_speed', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'vapor_pressure_deficit', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'vapor_pressure', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'snow_water_equivalent', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'shortwave_radiation_flux', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'soil_moisture', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'runoff', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'precipitation', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'potential_evaporation', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'climatic_water_deficit', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'water_evaporation', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss',
+    'TerraClimate', 'palmer_drought_severity_index', NA, '1958-2020', 'Municipality', 'http://thredds.northwestknowledge.net:8080/thredds/ncss'
   )
 
   return(link)
